Update signup and login according to design
